+++ packages/pangea-sdk/PangeaCyber.Net/BaseClient.cs
@@ -200,7 +200,6 @@ namespace PangeaCyber.Net
         ///
         protected async Task UploadPresignedURL(string url, TransferMethod transferMethod, FileData fileData)
         {
-
             try
             {
                 HttpResponseMessage resPSurl;