+++ packages/pangea-sdk/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Documentation for client builders and constructors.
 - Audit /download_results endpoint support
+- Sanitize service support
 
 ### Fixed